JENKINS-49337 add backward compatibility
